fix problems when not passing in role and disease

+++ remotemed_api/blueprints/users/views.py
@@ -147,53 +147,56 @@ def edit():
             user.guardian = new_guardian.id
         else:
             pass
-
-
         role = request.json.get('role')
         disease = request.json.get('disease')
 
         if user.save():
             if user.id != online_user['id']:
-                
-                role_list = UserRole.select().where(UserRole.user_id == user.id) #select all existing role(s)
-                #Delete obselete role
-                for r in role_list:
-                    if r not in role:
-                        del_role = UserRole.get_or_none(UserRole.role_id == r.role_id, UserRole.user_id == user.id)
-                        del_role.delete_instance()
-                #Add new role
-                for i in range(len(role)):
-                    if role[i-1] not in role_list:
-                        new_role = UserRole(user = user, role_id = role[i-1])
-                        if new_role.save():
-                            pass
-                        else:
-                            response = {
-                                    "message": "Can't add new role, please try again",
-                                    "status": "failed"
-                                }
-                if "1" in role:
-                    disease_list = UserDisease.select().where(UserDisease.user_id == user.id) #select all existing disease(s)
-                    #Add new disease
-                    for d in disease_list:
-                        if d not in disease:
-                            del_disease = UserDisease.get_or_none(UserDisease.disease_id == d.disease_id, UserDisease.user_id == user.id)
-                            del_disease.delete_instance()                        
-                    #delete obsolete disease
-                    for i in range(len(disease)):
-                        if disease[i-1] not in disease_list:
-                            new_disease = UserDisease(user = user, disease_id = disease[i-1])
-                            if new_disease.save():
+                if role:
+                    role_list = UserRole.select().where(UserRole.user_id == user.id) #select all existing role(s)
+                    #Delete obselete role
+                    for r in role_list:
+                        if r not in role:
+                            del_role = UserRole.get_or_none(UserRole.role_id == r.role_id, UserRole.user_id == user.id)
+                            del_role.delete_instance()
+                    #Add new role
+                    for i in range(len(role)):
+                        if role[i-1] not in role_list:
+                            new_role = UserRole(user = user, role_id = role[i-1])
+                            if new_role.save():
                                 pass
+                            else:
+                                response = {
+                                        "message": "Can't add new role, please try again",
+                                        "status": "failed"
+                                    }
+                else:
+                    pass
+                    if "1" in role:
+                        if disease:
+                            disease_list = UserDisease.select().where(UserDisease.user_id == user.id) #select all existing disease(s)
+                            #Add new disease
+                            for d in disease_list:
+                                if d not in disease:
+                                    del_disease = UserDisease.get_or_none(UserDisease.disease_id == d.disease_id, UserDisease.user_id == user.id)
+                                    del_disease.delete_instance()                        
+                            #delete obsolete disease
+                            for i in range(len(disease)):
+                                if disease[i-1] not in disease_list:
+                                    new_disease = UserDisease(user = user, disease_id = disease[i-1])
+                                    if new_disease.save():
+                                        pass
+                                else:
+                                    response = {
+                                            "message": "Can't add new disease, please try again",
+                                            "status": "failed"
+                                        }
                         else:
-                            response = {
-                                    "message": "Can't add new disease, please try again",
-                                    "status": "failed"
-                                }
-                response = {
-                    "message": "Updated user info successfully!",
-                    "status": "success"
-                }
+                            pass
+                    response = {
+                        "message": "Updated user info successfully!",
+                        "status": "success"
+                    }
             else:
                 response = {
                     "message": "Updated user info successfully!",
